Moved MockJsRuntime to its own namespace
