+++ src/Mocking/JSInterop/MockJsRuntimeInvokeHandler.cs
@@ -6,7 +6,7 @@ using Xunit;
 using System;
 using System.Linq;
 
-namespace Egil.RazorComponents.Testing
+namespace Egil.RazorComponents.Testing.Mocking.JSInterop
 {
     /// <summary>
     /// Represents an invoke handler for a mock of a <see cref="IJSRuntime"/>.
